Feat: 전체 사진 목록 페이지 로딩 및 에러 핸들링
